combined getting of datasets into one function, also cleaned up kitti initialization a bit
